+++ view/index.js
@@ -110,7 +110,7 @@ function getCurrentTreePosition() {
 }
 
 function setCurrentTreePosition(tree, index, now, redraw) {
-    if (!tree || getScoringMode()) return
+    if (!tree || getScoringMode() || getEstimatorMode()) return
 
     // Remove old graph node color
 
@@ -1301,16 +1301,19 @@ function findMove(vertex, text, step) {
 function vertexClicked(vertex, event) {
     closeGameInfo()
 
-    if (getScoringMode()) {
+    if (getScoringMode() || getEstimatorMode()) {
         if ($('score').hasClass('show')) return
         if (event.button != 0) return
         if (getBoard().arrangement[vertex] == 0) return
 
-        getBoard().getRelatedChains(vertex).forEach(function(v) {
-            $$('#goban .pos_' + v[0] + '-' + v[1]).toggleClass('dead')
+        var dead = !$$('#goban .pos_' + vertex.join('-'))[0].hasClass('dead')
+        var stones = getEstimatorMode() ? getBoard().getChain(vertex) : getBoard().getRelatedChains(vertex)
+
+        stones.forEach(function(v) {
+            $$('#goban .pos_' + v.join('-')).toggleClass('dead', dead)
         })
 
-        updateAreaMap()
+        updateAreaMap(getEstimatorMode())
     } else if (getEditMode()) {
         useTool(vertex, event)
     } else if (getFindMode()) {
@@ -1451,7 +1454,7 @@ function updateCommentText() {
     $('properties').retrieve('scrollbar').update()
 }
 
-function updateAreaMap() {
+function updateAreaMap(useEstimateMap) {
     var board = getBoard().clone()
 
     $$('#goban .row li.dead').forEach(function(li) {
@@ -1461,18 +1464,20 @@ function updateAreaMap() {
         board.arrangement[li.retrieve('vertex')] = 0
     })
 
-    var map = board.getAreaMap()
+    var map = useEstimateMap ? board.getAreaEstimateMap() : board.getAreaMap()
 
     $$('#goban .row li').forEach(function(li) {
         li.removeClass('area_-1').removeClass('area_0').removeClass('area_1')
             .addClass('area_' + map[li.retrieve('vertex')])
     })
 
-    var falsedead = $$('#goban .row li.area_-1.sign_-1.dead, #goban .row li.area_1.sign_1.dead')
+    if (!useEstimateMap) {
+        var falsedead = $$('#goban .row li.area_-1.sign_-1.dead, #goban .row li.area_1.sign_1.dead')
 
-    if (falsedead.length > 0) {
-        falsedead.removeClass('dead')
-        return updateAreaMap()
+        if (falsedead.length > 0) {
+            falsedead.removeClass('dead')
+            return updateAreaMap()
+        }
     }
 
     $('goban').store('areamap', map)